a bit of refactor
